Convert event_subtype to 32 bits. Add error module.

+++ include/bandsel.hpp
@@ -17,14 +17,14 @@ class BANDSEL
 {
   public:
     // Initialization function.
-    bool begin(class PCA9554 *p_bpf, class PCA9554 *p_lpf, void (*event_callback)(uint32_t, uint8_t, event_data), BANDS p_init_band = BAND_20M );
+    bool begin(class PCA9554 *p_bpf, class PCA9554 *p_lpf, void (*event_callback)(uint32_t, uint32_t, event_data), BANDS p_init_band = BAND_20M );
     
     // Set the desired band for BPF and LPF.
     void set(BANDS new_band);
   
   private:
     void fire_event(uint32_t event_type, uint32_t event_subtype);
-    void (*ev_cb)(uint32_t, uint8_t, event_data);
+    void (*ev_cb)(uint32_t, uint32_t, event_data);
     class PCA9554 *bpf;
     class PCA9554 *lpf;
