+++ include/encoder.hpp
@@ -6,17 +6,17 @@ class ENCODER
     public:
     void begin(uint8_t gpio_i, uint8_t gpio_q, uint8_t gpio_switch, 
         void (*interrupt_callback)() = NULL, 
-        void (*event_callback)(uint32_t event, uint8_t event_subtype, event_data ed) = NULL,
+        void (*event_callback)(uint32_t event, uint32_t event_subtype, event_data ed) = NULL,
         bool invert_encoder_signals = true, bool invert_switch_signal = true);
    
     void interrupt_handler();
-    void handler(event_data ed, uint8_t event_subtype);
+    void handler(event_data ed, uint32_t event_subtype);
    
     private:
 
     void poll();
     void (*interrupt_function)();
-    void (*callback_function)(uint32_t event, uint8_t event_subtype, event_data ed);
+    void (*callback_function)(uint32_t event, uint32_t event_subtype, event_data ed);
     bool phase_invert;
     bool switch_invert;
